refactor: split monolithic database.py into modular structure
