+++ backend/alembic/env.py
@@ -1,9 +1,9 @@
 from logging.config import fileConfig
 
-from app import Base
 from sqlalchemy import engine_from_config, pool
 
 from alembic import context
+from src.database import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -64,9 +64,7 @@ def run_migrations_online() -> None:
     )
 
     with connectable.connect() as connection:
-        context.configure(
-            connection=connection, target_metadata=target_metadata
-        )
+        context.configure(connection=connection, target_metadata=target_metadata)
 
         with context.begin_transaction():
             context.run_migrations()
