+++ backend/src/repositories/__init__.py
@@ -1,0 +1,3 @@
+from .news_repository import NewsArticleRepository
+
+__all__ = ["NewsArticleRepository"]
